split smoketests
