+++ crates/smoketests/tests/smoketests/cli/mod.rs
@@ -1,6 +1,5 @@
 pub mod auth;
 pub mod dev;
 pub mod generate;
-pub mod list;
 pub mod publish;
 pub mod server;
