+++ crates/smoketests/tests/smoketests/views.rs
@@ -1,12 +1,5 @@
-use std::path::PathBuf;
-
 use serde_json::{json, Value};
-use spacetimedb_smoketests::{
-    random_string, require_dotnet, require_local_server, require_pnpm, workspace_root, ModuleLanguage, Smoketest,
-};
-
-const STALE_VIEW_BACKING_TABLE_FIXTURE_IDENTITY: &str =
-    "c200f6ec405075e508c2ed6474019332d6a2a46c69614306cc4bd980e0b8b767";
+use spacetimedb_smoketests::{random_string, require_dotnet, require_pnpm, ModuleLanguage, Smoketest};
 
 const TS_VIEWS_SUBSCRIBE_MODULE: &str = r#"import { schema, t, table } from "spacetimedb/server";
 
@@ -223,27 +216,6 @@ fn project_fields(events: Vec<Value>, view_name: &str, projected_fields: &[&str]
             })
         })
         .collect()
-}
-
-fn stale_view_backing_table_fixture() -> PathBuf {
-    workspace_root()
-        .join("crates")
-        .join("smoketests")
-        .join("fixtures")
-        .join("stale-view-backing-table-v2.6.0")
-}
-
-fn stale_view_backing_table_test() -> Smoketest {
-    let test = Smoketest::builder()
-        .data_dir_fixture(
-            stale_view_backing_table_fixture(),
-            STALE_VIEW_BACKING_TABLE_FIXTURE_IDENTITY,
-        )
-        .autopublish(false)
-        .build();
-
-    test.new_identity().unwrap();
-    test
 }
 
 fn assert_count_view_refresh_behavior(test: &Smoketest, view_name: &str, id: &str, value: &str, updated_value: &str) {
@@ -645,78 +617,6 @@ fn test_view_primary_key_auto_migration_disconnects_clients() {
             .any(|l| l.contains("VIEW PRIMARY KEY UPDATE: client disconnected")),
         "Expected client_disconnected reducer log in logs: {:?}",
         logs
-    );
-}
-
-#[test]
-fn test_repair_stale_sender_scoped_view_backing_table_on_startup() {
-    require_local_server!();
-
-    let test = stale_view_backing_table_test();
-
-    let sender_view_sub = test
-        .subscribe(&["select * from player"])
-        .expect_rows(2)
-        .background()
-        .unwrap();
-
-    test.call("set_player_state", &["42", "1"]).unwrap();
-    test.call("set_player_state", &["42", "2"]).unwrap();
-
-    let sender_view_events = sender_view_sub.collect().unwrap();
-    let sender_view_projection = project_fields(sender_view_events, "player", &["id", "level"]);
-    assert_eq!(
-        serde_json::json!(sender_view_projection),
-        json!([
-            {
-                "player": {
-                    "deletes": [],
-                    "inserts": [{ "id": 42, "level": 1 }]
-                }
-            },
-            {
-                "player": {
-                    "deletes": [{ "id": 42, "level": 1 }],
-                    "inserts": [{ "id": 42, "level": 2 }]
-                }
-            }
-        ])
-    );
-}
-
-#[test]
-fn test_repair_stale_anonymous_view_backing_table_on_startup() {
-    require_local_server!();
-
-    let test = stale_view_backing_table_test();
-
-    let anonymous_view_sub = test
-        .subscribe(&["select * from player_and_level"])
-        .expect_rows(2)
-        .background()
-        .unwrap();
-
-    test.call("add_player_level", &["1", "2"]).unwrap();
-    test.call("add_player_level", &["2", "2"]).unwrap();
-
-    let anonymous_view_events = anonymous_view_sub.collect().unwrap();
-    let anonymous_view_projection = project_fields(anonymous_view_events, "player_and_level", &["id", "level"]);
-    assert_eq!(
-        serde_json::json!(anonymous_view_projection),
-        json!([
-            {
-                "player_and_level": {
-                    "deletes": [],
-                    "inserts": [{ "id": 1, "level": 2 }]
-                }
-            },
-            {
-                "player_and_level": {
-                    "deletes": [],
-                    "inserts": [{ "id": 2, "level": 2 }]
-                }
-            }
-        ])
     );
 }
