+++ crates/smoketests/tests/standalone/cli/mod.rs
@@ -1,0 +1,4 @@
+#[path = "../../smoketests/cli/list.rs"]
+mod list;
+#[path = "../cli_server.rs"]
+mod server;
