+++ crates/smoketests/tests/standalone/servers.rs
@@ -1,0 +1,67 @@
+use regex::Regex;
+use spacetimedb_smoketests::{require_local_server, Smoketest};
+
+/// Verify that we can add and list server configurations.
+#[test]
+fn test_servers() {
+    // This only covers local CLI config behavior, so it is not valuable to run
+    // against remote servers.
+    require_local_server!();
+
+    let test = Smoketest::builder().autopublish(false).build();
+
+    // Add a test server (local-only command, no --server flag needed)
+    let output = test
+        .spacetime(&[
+            "server",
+            "add",
+            "--url",
+            "https://testnet.spacetimedb.com",
+            "testnet",
+            "--no-fingerprint",
+        ])
+        .unwrap();
+
+    assert!(
+        Regex::new(r"Host: testnet\.spacetimedb\.com\n")
+            .unwrap()
+            .is_match(&output),
+        "Expected host in output: {}",
+        output
+    );
+    assert!(
+        Regex::new(r"Protocol: https\n").unwrap().is_match(&output),
+        "Expected host in output: {}",
+        output
+    );
+
+    // List servers (local-only command)
+    let servers = test.spacetime(&["server", "list"]).unwrap();
+
+    let testnet_re = Regex::new(r"(?m)^\s*testnet\.spacetimedb\.com\s+https\s+testnet\s*$").unwrap();
+    assert!(
+        testnet_re.is_match(&servers),
+        "Expected testnet in server list: {}",
+        servers
+    );
+
+    // Add the local test server to the config so we can check its fingerprint
+    test.spacetime(&[
+        "server",
+        "add",
+        "--url",
+        &test.server_url,
+        "test-local",
+        "--no-fingerprint",
+    ])
+    .unwrap();
+
+    // Check fingerprint commands (local-only command)
+    let output = test.spacetime(&["server", "fingerprint", "test-local", "-y"]).unwrap();
+    // The exact message may vary, just check it doesn't error
+    assert!(
+        output.contains("fingerprint") || output.contains("Fingerprint"),
+        "Expected fingerprint message: {}",
+        output
+    );
+}
